LayoutTests/platform/qt/fast/js/global-constructors-expected.txt

This test documents our set of global constructors we expose on the window object (FF and Opera don't expose them on the window, btw). This also checks to make sure than any constructor attribute we expose has the expected constructor type.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Attr.toString() is '[object AttrConstructor]'
PASS Audio.toString() is '[object AudioConstructor]'
PASS BeforeLoadEvent.toString() is '[object BeforeLoadEventConstructor]'
PASS Blob.toString() is '[object BlobConstructor]'
PASS CDATASection.toString() is '[object CDATASectionConstructor]'
PASS CSSCharsetRule.toString() is '[object CSSCharsetRuleConstructor]'
PASS CSSFontFaceRule.toString() is '[object CSSFontFaceRuleConstructor]'
PASS CSSImportRule.toString() is '[object CSSImportRuleConstructor]'
PASS CSSMediaRule.toString() is '[object CSSMediaRuleConstructor]'
PASS CSSPageRule.toString() is '[object CSSPageRuleConstructor]'
PASS CSSPrimitiveValue.toString() is '[object CSSPrimitiveValueConstructor]'
PASS CSSRule.toString() is '[object CSSRuleConstructor]'
PASS CSSRuleList.toString() is '[object CSSRuleListConstructor]'
PASS CSSStyleDeclaration.toString() is '[object CSSStyleDeclarationConstructor]'
PASS CSSStyleRule.toString() is '[object CSSStyleRuleConstructor]'
PASS CSSStyleSheet.toString() is '[object CSSStyleSheetConstructor]'
PASS CSSValue.toString() is '[object CSSValueConstructor]'
PASS CSSValueList.toString() is '[object CSSValueListConstructor]'
PASS CanvasGradient.toString() is '[object CanvasGradientConstructor]'
PASS CanvasPattern.toString() is '[object CanvasPatternConstructor]'
PASS CanvasRenderingContext2D.toString() is '[object CanvasRenderingContext2DConstructor]'
PASS CharacterData.toString() is '[object CharacterDataConstructor]'
PASS ClientRect.toString() is '[object ClientRectConstructor]'
PASS ClientRectList.toString() is '[object ClientRectListConstructor]'
PASS Clipboard.toString() is '[object ClipboardConstructor]'
PASS CloseEvent.toString() is '[object CloseEventConstructor]'
PASS Comment.toString() is '[object CommentConstructor]'
PASS CompositionEvent.toString() is '[object CompositionEventConstructor]'
PASS Counter.toString() is '[object CounterConstructor]'
PASS CustomEvent.toString() is '[object CustomEventConstructor]'
PASS DOMException.toString() is '[object DOMExceptionConstructor]'
PASS DOMImplementation.toString() is '[object DOMImplementationConstructor]'
PASS DOMParser.toString() is '[object DOMParserConstructor]'
PASS DOMSettableTokenList.toString() is '[object DOMSettableTokenListConstructor]'
PASS DOMStringList.toString() is '[object DOMStringListConstructor]'
PASS DOMStringMap.toString() is '[object DOMStringMapConstructor]'
PASS DOMTokenList.toString() is '[object DOMTokenListConstructor]'
PASS DeviceMotionEvent.toString() is '[object DeviceMotionEventConstructor]'
PASS DeviceOrientationEvent.toString() is '[object DeviceOrientationEventConstructor]'
PASS Document.toString() is '[object DocumentConstructor]'
PASS DocumentFragment.toString() is '[object DocumentFragmentConstructor]'
PASS DocumentType.toString() is '[object DocumentTypeConstructor]'
PASS Element.toString() is '[object ElementConstructor]'
PASS Entity.toString() is '[object EntityConstructor]'
PASS EntityReference.toString() is '[object EntityReferenceConstructor]'
PASS ErrorEvent.toString() is '[object ErrorEventConstructor]'
PASS Event.toString() is '[object EventConstructor]'
PASS EventException.toString() is '[object EventExceptionConstructor]'
PASS EventSource.toString() is '[object EventSourceConstructor]'
PASS File.toString() is '[object FileConstructor]'
PASS FileList.toString() is '[object FileListConstructor]'
PASS FormData.toString() is '[object FormDataConstructor]'
PASS HTMLAllCollection.toString() is '[object HTMLAllCollectionConstructor]'
PASS HTMLAnchorElement.toString() is '[object HTMLAnchorElementConstructor]'
PASS HTMLAppletElement.toString() is '[object HTMLAppletElementConstructor]'
PASS HTMLAreaElement.toString() is '[object HTMLAreaElementConstructor]'
PASS HTMLAudioElement.toString() is '[object HTMLAudioElementConstructor]'
PASS HTMLBRElement.toString() is '[object HTMLBRElementConstructor]'
PASS HTMLBaseElement.toString() is '[object HTMLBaseElementConstructor]'
PASS HTMLBaseFontElement.toString() is '[object HTMLBaseFontElementConstructor]'
PASS HTMLBodyElement.toString() is '[object HTMLBodyElementConstructor]'
PASS HTMLButtonElement.toString() is '[object HTMLButtonElementConstructor]'
PASS HTMLCanvasElement.toString() is '[object HTMLCanvasElementConstructor]'
PASS HTMLCollection.toString() is '[object HTMLCollectionConstructor]'
PASS HTMLDListElement.toString() is '[object HTMLDListElementConstructor]'
PASS HTMLDirectoryElement.toString() is '[object HTMLDirectoryElementConstructor]'
PASS HTMLDivElement.toString() is '[object HTMLDivElementConstructor]'
PASS HTMLDocument.toString() is '[object HTMLDocumentConstructor]'
PASS HTMLElement.toString() is '[object HTMLElementConstructor]'
PASS HTMLEmbedElement.toString() is '[object HTMLEmbedElementConstructor]'
PASS HTMLFieldSetElement.toString() is '[object HTMLFieldSetElementConstructor]'
PASS HTMLFontElement.toString() is '[object HTMLFontElementConstructor]'
PASS HTMLFormElement.toString() is '[object HTMLFormElementConstructor]'
PASS HTMLFrameElement.toString() is '[object HTMLFrameElementConstructor]'
PASS HTMLFrameSetElement.toString() is '[object HTMLFrameSetElementConstructor]'
PASS HTMLHRElement.toString() is '[object HTMLHRElementConstructor]'
PASS HTMLHeadElement.toString() is '[object HTMLHeadElementConstructor]'
PASS HTMLHeadingElement.toString() is '[object HTMLHeadingElementConstructor]'
PASS HTMLHtmlElement.toString() is '[object HTMLHtmlElementConstructor]'
PASS HTMLIFrameElement.toString() is '[object HTMLIFrameElementConstructor]'
PASS HTMLImageElement.toString() is '[object HTMLImageElementConstructor]'
PASS HTMLInputElement.toString() is '[object HTMLInputElementConstructor]'
PASS HTMLKeygenElement.toString() is '[object HTMLKeygenElementConstructor]'
PASS HTMLLIElement.toString() is '[object HTMLLIElementConstructor]'
PASS HTMLLabelElement.toString() is '[object HTMLLabelElementConstructor]'
PASS HTMLLegendElement.toString() is '[object HTMLLegendElementConstructor]'
PASS HTMLLinkElement.toString() is '[object HTMLLinkElementConstructor]'
PASS HTMLMapElement.toString() is '[object HTMLMapElementConstructor]'
PASS HTMLMarqueeElement.toString() is '[object HTMLMarqueeElementConstructor]'
PASS HTMLMediaElement.toString() is '[object HTMLMediaElementConstructor]'
PASS HTMLMenuElement.toString() is '[object HTMLMenuElementConstructor]'
PASS HTMLMetaElement.toString() is '[object HTMLMetaElementConstructor]'
PASS HTMLMeterElement.toString() is '[object HTMLMeterElementConstructor]'
PASS HTMLModElement.toString() is '[object HTMLModElementConstructor]'
PASS HTMLOListElement.toString() is '[object HTMLOListElementConstructor]'
PASS HTMLObjectElement.toString() is '[object HTMLObjectElementConstructor]'
PASS HTMLOptGroupElement.toString() is '[object HTMLOptGroupElementConstructor]'
PASS HTMLOptionElement.toString() is '[object HTMLOptionElementConstructor]'
PASS HTMLOutputElement.toString() is '[object HTMLOutputElementConstructor]'
PASS HTMLParagraphElement.toString() is '[object HTMLParagraphElementConstructor]'
PASS HTMLParamElement.toString() is '[object HTMLParamElementConstructor]'
PASS HTMLPreElement.toString() is '[object HTMLPreElementConstructor]'
PASS HTMLProgressElement.toString() is '[object HTMLProgressElementConstructor]'
PASS HTMLQuoteElement.toString() is '[object HTMLQuoteElementConstructor]'
PASS HTMLScriptElement.toString() is '[object HTMLScriptElementConstructor]'
PASS HTMLSelectElement.toString() is '[object HTMLSelectElementConstructor]'
PASS HTMLSourceElement.toString() is '[object HTMLSourceElementConstructor]'
PASS HTMLSpanElement.toString() is '[object HTMLSpanElementConstructor]'
PASS HTMLStyleElement.toString() is '[object HTMLStyleElementConstructor]'
PASS HTMLTableCaptionElement.toString() is '[object HTMLTableCaptionElementConstructor]'
PASS HTMLTableCellElement.toString() is '[object HTMLTableCellElementConstructor]'
PASS HTMLTableColElement.toString() is '[object HTMLTableColElementConstructor]'
PASS HTMLTableElement.toString() is '[object HTMLTableElementConstructor]'
PASS HTMLTableRowElement.toString() is '[object HTMLTableRowElementConstructor]'
PASS HTMLTableSectionElement.toString() is '[object HTMLTableSectionElementConstructor]'
PASS HTMLTextAreaElement.toString() is '[object HTMLTextAreaElementConstructor]'
PASS HTMLTitleElement.toString() is '[object HTMLTitleElementConstructor]'
PASS HTMLUListElement.toString() is '[object HTMLUListElementConstructor]'
PASS HTMLUnknownElement.toString() is '[object HTMLUnknownElementConstructor]'
PASS HTMLVideoElement.toString() is '[object HTMLVideoElementConstructor]'
PASS HashChangeEvent.toString() is '[object HashChangeEventConstructor]'
PASS Image.toString() is '[object ImageConstructor]'
PASS ImageData.toString() is '[object ImageDataConstructor]'
PASS KeyboardEvent.toString() is '[object KeyboardEventConstructor]'
PASS MediaController.toString() is '[object MediaControllerConstructor]'
PASS MediaError.toString() is '[object MediaErrorConstructor]'
PASS MediaList.toString() is '[object MediaListConstructor]'
PASS MessageChannel.toString() is '[object MessageChannelConstructor]'
PASS MessageEvent.toString() is '[object MessageEventConstructor]'
PASS MessagePort.toString() is '[object MessagePortConstructor]'
PASS MimeType.toString() is '[object MimeTypeConstructor]'
PASS MimeTypeArray.toString() is '[object MimeTypeArrayConstructor]'
PASS MouseEvent.toString() is '[object MouseEventConstructor]'
PASS MutationEvent.toString() is '[object MutationEventConstructor]'
PASS NamedNodeMap.toString() is '[object NamedNodeMapConstructor]'
PASS Node.toString() is '[object NodeConstructor]'
PASS NodeFilter.toString() is '[object NodeFilterConstructor]'
PASS NodeList.toString() is '[object NodeListConstructor]'
PASS Notation.toString() is '[object NotationConstructor]'
PASS Option.toString() is '[object OptionConstructor]'
PASS OverflowEvent.toString() is '[object OverflowEventConstructor]'
PASS PageTransitionEvent.toString() is '[object PageTransitionEventConstructor]'
PASS Plugin.toString() is '[object PluginConstructor]'
PASS PluginArray.toString() is '[object PluginArrayConstructor]'
PASS PopStateEvent.toString() is '[object PopStateEventConstructor]'
PASS ProcessingInstruction.toString() is '[object ProcessingInstructionConstructor]'
PASS ProgressEvent.toString() is '[object ProgressEventConstructor]'
PASS RGBColor.toString() is '[object RGBColorConstructor]'
PASS Range.toString() is '[object RangeConstructor]'
PASS RangeException.toString() is '[object RangeExceptionConstructor]'
PASS Rect.toString() is '[object RectConstructor]'
PASS SQLException.toString() is '[object SQLExceptionConstructor]'
PASS SVGAElement.toString() is '[object SVGAElementConstructor]'
PASS SVGAngle.toString() is '[object SVGAngleConstructor]'
PASS SVGAnimateColorElement.toString() is '[object SVGAnimateColorElementConstructor]'
PASS SVGAnimateElement.toString() is '[object SVGAnimateElementConstructor]'
PASS SVGAnimateMotionElement.toString() is '[object SVGAnimateMotionElementConstructor]'
PASS SVGAnimateTransformElement.toString() is '[object SVGAnimateTransformElementConstructor]'
PASS SVGAnimatedAngle.toString() is '[object SVGAnimatedAngleConstructor]'
PASS SVGAnimatedBoolean.toString() is '[object SVGAnimatedBooleanConstructor]'
PASS SVGAnimatedEnumeration.toString() is '[object SVGAnimatedEnumerationConstructor]'
PASS SVGAnimatedInteger.toString() is '[object SVGAnimatedIntegerConstructor]'
PASS SVGAnimatedLength.toString() is '[object SVGAnimatedLengthConstructor]'
PASS SVGAnimatedLengthList.toString() is '[object SVGAnimatedLengthListConstructor]'
PASS SVGAnimatedNumber.toString() is '[object SVGAnimatedNumberConstructor]'
PASS SVGAnimatedNumberList.toString() is '[object SVGAnimatedNumberListConstructor]'
PASS SVGAnimatedPreserveAspectRatio.toString() is '[object SVGAnimatedPreserveAspectRatioConstructor]'
PASS SVGAnimatedRect.toString() is '[object SVGAnimatedRectConstructor]'
PASS SVGAnimatedString.toString() is '[object SVGAnimatedStringConstructor]'
PASS SVGAnimatedTransformList.toString() is '[object SVGAnimatedTransformListConstructor]'
PASS SVGCircleElement.toString() is '[object SVGCircleElementConstructor]'
PASS SVGClipPathElement.toString() is '[object SVGClipPathElementConstructor]'
PASS SVGColor.toString() is '[object SVGColorConstructor]'
PASS SVGComponentTransferFunctionElement.toString() is '[object SVGComponentTransferFunctionElementConstructor]'
PASS SVGCursorElement.toString() is '[object SVGCursorElementConstructor]'
PASS SVGDefsElement.toString() is '[object SVGDefsElementConstructor]'
PASS SVGDescElement.toString() is '[object SVGDescElementConstructor]'
PASS SVGDocument.toString() is '[object SVGDocumentConstructor]'
PASS SVGElement.toString() is '[object SVGElementConstructor]'
PASS SVGElementInstance.toString() is '[object SVGElementInstanceConstructor]'
PASS SVGElementInstanceList.toString() is '[object SVGElementInstanceListConstructor]'
PASS SVGEllipseElement.toString() is '[object SVGEllipseElementConstructor]'
PASS SVGException.toString() is '[object SVGExceptionConstructor]'
PASS SVGFEBlendElement.toString() is '[object SVGFEBlendElementConstructor]'
PASS SVGFEColorMatrixElement.toString() is '[object SVGFEColorMatrixElementConstructor]'
PASS SVGFEComponentTransferElement.toString() is '[object SVGFEComponentTransferElementConstructor]'
PASS SVGFECompositeElement.toString() is '[object SVGFECompositeElementConstructor]'
PASS SVGFEConvolveMatrixElement.toString() is '[object SVGFEConvolveMatrixElementConstructor]'
PASS SVGFEDiffuseLightingElement.toString() is '[object SVGFEDiffuseLightingElementConstructor]'
PASS SVGFEDisplacementMapElement.toString() is '[object SVGFEDisplacementMapElementConstructor]'
PASS SVGFEDistantLightElement.toString() is '[object SVGFEDistantLightElementConstructor]'
PASS SVGFEDropShadowElement.toString() is '[object SVGFEDropShadowElementConstructor]'
PASS SVGFEFloodElement.toString() is '[object SVGFEFloodElementConstructor]'
PASS SVGFEFuncAElement.toString() is '[object SVGFEFuncAElementConstructor]'
PASS SVGFEFuncBElement.toString() is '[object SVGFEFuncBElementConstructor]'
PASS SVGFEFuncGElement.toString() is '[object SVGFEFuncGElementConstructor]'
PASS SVGFEFuncRElement.toString() is '[object SVGFEFuncRElementConstructor]'
PASS SVGFEGaussianBlurElement.toString() is '[object SVGFEGaussianBlurElementConstructor]'
PASS SVGFEImageElement.toString() is '[object SVGFEImageElementConstructor]'
PASS SVGFEMergeElement.toString() is '[object SVGFEMergeElementConstructor]'
PASS SVGFEMergeNodeElement.toString() is '[object SVGFEMergeNodeElementConstructor]'
PASS SVGFEMorphologyElement.toString() is '[object SVGFEMorphologyElementConstructor]'
PASS SVGFEOffsetElement.toString() is '[object SVGFEOffsetElementConstructor]'
PASS SVGFEPointLightElement.toString() is '[object SVGFEPointLightElementConstructor]'
PASS SVGFESpecularLightingElement.toString() is '[object SVGFESpecularLightingElementConstructor]'
PASS SVGFESpotLightElement.toString() is '[object SVGFESpotLightElementConstructor]'
PASS SVGFETileElement.toString() is '[object SVGFETileElementConstructor]'
PASS SVGFETurbulenceElement.toString() is '[object SVGFETurbulenceElementConstructor]'
PASS SVGFilterElement.toString() is '[object SVGFilterElementConstructor]'
PASS SVGForeignObjectElement.toString() is '[object SVGForeignObjectElementConstructor]'
PASS SVGGElement.toString() is '[object SVGGElementConstructor]'
PASS SVGGradientElement.toString() is '[object SVGGradientElementConstructor]'
PASS SVGImageElement.toString() is '[object SVGImageElementConstructor]'
PASS SVGLength.toString() is '[object SVGLengthConstructor]'
PASS SVGLengthList.toString() is '[object SVGLengthListConstructor]'
PASS SVGLineElement.toString() is '[object SVGLineElementConstructor]'
PASS SVGLinearGradientElement.toString() is '[object SVGLinearGradientElementConstructor]'
PASS SVGMPathElement.toString() is '[object SVGMPathElementConstructor]'
PASS SVGMarkerElement.toString() is '[object SVGMarkerElementConstructor]'
PASS SVGMaskElement.toString() is '[object SVGMaskElementConstructor]'
PASS SVGMatrix.toString() is '[object SVGMatrixConstructor]'
PASS SVGMetadataElement.toString() is '[object SVGMetadataElementConstructor]'
PASS SVGNumber.toString() is '[object SVGNumberConstructor]'
PASS SVGNumberList.toString() is '[object SVGNumberListConstructor]'
PASS SVGPaint.toString() is '[object SVGPaintConstructor]'
PASS SVGPathElement.toString() is '[object SVGPathElementConstructor]'
PASS SVGPathSeg.toString() is '[object SVGPathSegConstructor]'
PASS SVGPathSegArcAbs.toString() is '[object SVGPathSegArcAbsConstructor]'
PASS SVGPathSegArcRel.toString() is '[object SVGPathSegArcRelConstructor]'
PASS SVGPathSegClosePath.toString() is '[object SVGPathSegClosePathConstructor]'
PASS SVGPathSegCurvetoCubicAbs.toString() is '[object SVGPathSegCurvetoCubicAbsConstructor]'
PASS SVGPathSegCurvetoCubicRel.toString() is '[object SVGPathSegCurvetoCubicRelConstructor]'
PASS SVGPathSegCurvetoCubicSmoothAbs.toString() is '[object SVGPathSegCurvetoCubicSmoothAbsConstructor]'
PASS SVGPathSegCurvetoCubicSmoothRel.toString() is '[object SVGPathSegCurvetoCubicSmoothRelConstructor]'
PASS SVGPathSegCurvetoQuadraticAbs.toString() is '[object SVGPathSegCurvetoQuadraticAbsConstructor]'
PASS SVGPathSegCurvetoQuadraticRel.toString() is '[object SVGPathSegCurvetoQuadraticRelConstructor]'
PASS SVGPathSegCurvetoQuadraticSmoothAbs.toString() is '[object SVGPathSegCurvetoQuadraticSmoothAbsConstructor]'
PASS SVGPathSegCurvetoQuadraticSmoothRel.toString() is '[object SVGPathSegCurvetoQuadraticSmoothRelConstructor]'
PASS SVGPathSegLinetoAbs.toString() is '[object SVGPathSegLinetoAbsConstructor]'
PASS SVGPathSegLinetoHorizontalAbs.toString() is '[object SVGPathSegLinetoHorizontalAbsConstructor]'
PASS SVGPathSegLinetoHorizontalRel.toString() is '[object SVGPathSegLinetoHorizontalRelConstructor]'
PASS SVGPathSegLinetoRel.toString() is '[object SVGPathSegLinetoRelConstructor]'
PASS SVGPathSegLinetoVerticalAbs.toString() is '[object SVGPathSegLinetoVerticalAbsConstructor]'
PASS SVGPathSegLinetoVerticalRel.toString() is '[object SVGPathSegLinetoVerticalRelConstructor]'
PASS SVGPathSegList.toString() is '[object SVGPathSegListConstructor]'
PASS SVGPathSegMovetoAbs.toString() is '[object SVGPathSegMovetoAbsConstructor]'
PASS SVGPathSegMovetoRel.toString() is '[object SVGPathSegMovetoRelConstructor]'
PASS SVGPatternElement.toString() is '[object SVGPatternElementConstructor]'
PASS SVGPoint.toString() is '[object SVGPointConstructor]'
PASS SVGPointList.toString() is '[object SVGPointListConstructor]'
PASS SVGPolygonElement.toString() is '[object SVGPolygonElementConstructor]'
PASS SVGPolylineElement.toString() is '[object SVGPolylineElementConstructor]'
PASS SVGPreserveAspectRatio.toString() is '[object SVGPreserveAspectRatioConstructor]'
PASS SVGRadialGradientElement.toString() is '[object SVGRadialGradientElementConstructor]'
PASS SVGRect.toString() is '[object SVGRectConstructor]'
PASS SVGRectElement.toString() is '[object SVGRectElementConstructor]'
PASS SVGRenderingIntent.toString() is '[object SVGRenderingIntentConstructor]'
PASS SVGSVGElement.toString() is '[object SVGSVGElementConstructor]'
PASS SVGScriptElement.toString() is '[object SVGScriptElementConstructor]'
PASS SVGSetElement.toString() is '[object SVGSetElementConstructor]'
PASS SVGStopElement.toString() is '[object SVGStopElementConstructor]'
PASS SVGStringList.toString() is '[object SVGStringListConstructor]'
PASS SVGStyleElement.toString() is '[object SVGStyleElementConstructor]'
PASS SVGSwitchElement.toString() is '[object SVGSwitchElementConstructor]'
PASS SVGSymbolElement.toString() is '[object SVGSymbolElementConstructor]'
PASS SVGTRefElement.toString() is '[object SVGTRefElementConstructor]'
PASS SVGTSpanElement.toString() is '[object SVGTSpanElementConstructor]'
PASS SVGTextContentElement.toString() is '[object SVGTextContentElementConstructor]'
PASS SVGTextElement.toString() is '[object SVGTextElementConstructor]'
PASS SVGTextPathElement.toString() is '[object SVGTextPathElementConstructor]'
PASS SVGTextPositioningElement.toString() is '[object SVGTextPositioningElementConstructor]'
PASS SVGTitleElement.toString() is '[object SVGTitleElementConstructor]'
PASS SVGTransform.toString() is '[object SVGTransformConstructor]'
PASS SVGTransformList.toString() is '[object SVGTransformListConstructor]'
PASS SVGUnitTypes.toString() is '[object SVGUnitTypesConstructor]'
PASS SVGUseElement.toString() is '[object SVGUseElementConstructor]'
PASS SVGViewElement.toString() is '[object SVGViewElementConstructor]'
PASS SVGZoomEvent.toString() is '[object SVGZoomEventConstructor]'
PASS SharedWorker.toString() is '[object SharedWorkerConstructor]'
PASS Storage.toString() is '[object StorageConstructor]'
PASS StorageEvent.toString() is '[object StorageEventConstructor]'
PASS StyleSheet.toString() is '[object StyleSheetConstructor]'
PASS StyleSheetList.toString() is '[object StyleSheetListConstructor]'
PASS Text.toString() is '[object TextConstructor]'
PASS TextEvent.toString() is '[object TextEventConstructor]'
PASS TextMetrics.toString() is '[object TextMetricsConstructor]'
PASS TimeRanges.toString() is '[object TimeRangesConstructor]'
PASS TouchEvent.toString() is '[object TouchEventConstructor]'
PASS UIEvent.toString() is '[object UIEventConstructor]'
PASS WebGLContextEvent.toString() is '[object WebGLContextEventConstructor]'
PASS WebKitAnimationEvent.toString() is '[object WebKitAnimationEventConstructor]'
PASS WebKitCSSFilterValue.toString() is '[object WebKitCSSFilterValueConstructor]'
PASS WebKitCSSKeyframeRule.toString() is '[object WebKitCSSKeyframeRuleConstructor]'
PASS WebKitCSSKeyframesRule.toString() is '[object WebKitCSSKeyframesRuleConstructor]'
PASS WebKitCSSMatrix.toString() is '[object WebKitCSSMatrixConstructor]'
PASS WebKitCSSRegionRule.toString() is '[object WebKitCSSRegionRuleConstructor]'
PASS WebKitCSSTransformValue.toString() is '[object WebKitCSSTransformValueConstructor]'
PASS WebKitPoint.toString() is '[object WebKitPointConstructor]'
PASS WebKitTransitionEvent.toString() is '[object WebKitTransitionEventConstructor]'
PASS WebSocket.toString() is '[object WebSocketConstructor]'
PASS WheelEvent.toString() is '[object WheelEventConstructor]'
PASS Worker.toString() is '[object WorkerConstructor]'
PASS XMLDocument.toString() is '[object DocumentConstructor]'
PASS XMLHttpRequest.toString() is '[object XMLHttpRequestConstructor]'
PASS XMLHttpRequestException.toString() is '[object XMLHttpRequestExceptionConstructor]'
PASS XMLHttpRequestProgressEvent.toString() is '[object XMLHttpRequestProgressEventConstructor]'
PASS XMLHttpRequestUpload.toString() is '[object XMLHttpRequestUploadConstructor]'
PASS XMLSerializer.toString() is '[object XMLSerializerConstructor]'
PASS XPathEvaluator.toString() is '[object XPathEvaluatorConstructor]'
PASS XPathException.toString() is '[object XPathExceptionConstructor]'
PASS XPathResult.toString() is '[object XPathResultConstructor]'
PASS XSLTProcessor.toString() is '[object XSLTProcessorConstructor]'
FAIL webkitURL.toString() should be [object webkitURLConstructor]. Was [object DOMURLConstructor].
PASS successfullyParsed is true

TEST COMPLETE
